Currently having an issue with loading in the large training images data set into the data model used for image classification in Cinder. That would be the final step.

+++ include/core/data_model.h
@@ -101,7 +101,7 @@ class DataModel {
    * @param image The 2D vector corresponding to an image.
    * @return The model's prediction.
    */
-  size_t ClassifyImage(std::vector<std::vector<bool>> image);
+  int ClassifyImage(std::vector<std::vector<bool>> image);
   
   /** Getters */
   size_t GetImageDimensions() const;
